Redesigned the server as a lib, A server will be built using it but this way you can use it with your own logic.
