--- IDEAS.md
@@ -1,1 +1,0 @@
-Use (Nagle's)[https://en.wikipedia.org/wiki/Nagle's_algorithm] algorithm to avoid performance loss:
