+++ README.md
@@ -1,10 +1,9 @@
-# Yggdrasil
+# Yggio
 
 [![codecov](https://codecov.io/gh/Supamiu/yggdrasil/branch/master/graph/badge.svg)](https://codecov.io/gh/Supamiu/yggdrasil)
 [![Build Status](https://travis-ci.org/Supamiu/yggdrasil.svg?branch=master)](https://travis-ci.org/Supamiu/yggdrasil)
 
-Another mmo server written in Java, trying to create a simple way to add mmo layer to your game.
+A simple open source lib to create and send packets easily on your MMO server. It builds an abstract layer on top of java.nio.
 
-External dependencies:
-
- - You need to have RabbitMQ installed and running, see `rabbit.properties` file to configure the RabbitMQ connection.
+Projects using it: 
+ - [Yggdrasil](https://github.com/supamiu/yggdrasil) (mmo server)
