+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.supamiu</groupId>
-    <artifactId>yggdrasil</artifactId>
+    <artifactId>yggio</artifactId>
     <version>1.0-SNAPSHOT</version>
     <build>
         <plugins>
