--- src/main/resources/rabbitmq.properties
@@ -1,5 +1,0 @@
-username=guest
-password=guest
-vhost=/
-host=localhost
-port=5672
