+++ src/test/java/io/ygg/tcp/TcpProcessorTest.java
@@ -1,0 +1,9 @@
+package io.ygg.tcp;
+
+/**
+ * Created by flavien on 12/01/17.
+ */
+public class TcpProcessorTest {
+
+    //TODO At the moment I didn't find a way to test the TcpProcessor, maybe a bit of refactoring is needed.
+}
