--- src/test/java/net/yggdrasil/tcp/TcpProcessorTest.java
@@ -1,9 +1,0 @@
-package net.yggdrasil.tcp;
-
-/**
- * Created by flavien on 12/01/17.
- */
-public class TcpProcessorTest {
-
-    //TODO At the moment I didn't find a way to test the TcpProcessor, maybe a bit of refactoring is needed.
-}
